backend/src/requirements.txt
ultralytics
fastapi
SQLAlchemy
aiosqlite
aiortc
aiosqlite
alembic
a2wsgi 
python-dotenv
websockets